+++ packet.go
@@ -3,6 +3,7 @@ package rcon
 import (
 	"bytes"
 	"encoding/binary"
+	"fmt"
 	"io"
 )
 
@@ -91,7 +92,7 @@ func (packet *Packet) WriteTo(w io.Writer) (n int64, err error) {
 // ReadFrom implements io.ReaderFrom for read a packet from r.
 func (packet *Packet) ReadFrom(r io.Reader) (n int64, err error) {
 	if err := binary.Read(r, binary.LittleEndian, &packet.Size); err != nil {
-		return n, err
+		return n, fmt.Errorf("rcon: read packet size %w", err)
 	}
 
 	n += 4
@@ -101,13 +102,13 @@ func (packet *Packet) ReadFrom(r io.Reader) (n int64, err error) {
 	}
 
 	if err := binary.Read(r, binary.LittleEndian, &packet.ID); err != nil {
-		return n, err
+		return n, fmt.Errorf("rcon: read packet id %w", err)
 	}
 
 	n += 4
 
 	if err := binary.Read(r, binary.LittleEndian, &packet.Type); err != nil {
-		return n, err
+		return n, fmt.Errorf("rcon: read packet type: %w", err)
 	}
 
 	n += 4
@@ -121,7 +122,7 @@ func (packet *Packet) ReadFrom(r io.Reader) (n int64, err error) {
 		var m int
 
 		if m, err = r.Read(packet.body[i:]); err != nil {
-			return n + int64(m) + int64(i), err
+			return n + int64(m) + int64(i), fmt.Errorf("rcon: %w", err)
 		}
 
 		i += int32(m)