+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v1.3.1] - 2021-01-06
+### Updated
+- Updated golangci linter to 1.33 version
+
+### Changed
+- Changed errors handling - added wrapping.
+
 ## [v1.3.0] - 2020-12-02
 ### Fixed
 - Fixed wrong number of bytes written in Packet WriteTo function.
@@ -67,7 +74,8 @@ changed.
 ### Added
 - Initial implementation.
 
-[Unreleased]: https://github.com/gorcon/rcon/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/gorcon/rcon/compare/v1.3.1...HEAD
+[v1.3.1]: https://github.com/gorcon/rcon/compare/v1.3.0...v1.3.1
 [v1.3.0]: https://github.com/gorcon/rcon/compare/v1.2.4...v1.3.0
 [v1.2.4]: https://github.com/gorcon/rcon/compare/v1.2.3...v1.2.4
 [v1.2.3]: https://github.com/gorcon/rcon/compare/v1.2.2...v1.2.3